requirements.txt
requests
pillow
python-telegram-bot==13.15
beautifulsoup4
